added classification report to check class collapse and num_workers restored

+++ src/datasets_/objectnet.py
@@ -113,7 +113,7 @@ class ObjectNetBase(ImageNet):
         if logits.shape[1] == 113:
             return logits
         if torch.is_tensor(logits):
-            logits = logits.cpu().numpy()
+            logits = logits.float().cpu().numpy()
         logits_projected = np.zeros((logits.shape[0], 113))
         for k, v in self.rev_class_idx_map.items():
             logits_projected[:, k] = np.max(logits[:, v], axis=1).squeeze()
